requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
openai==1.40.0
pyyaml==6.0.2
httpx==0.27.0
